Do some type shifting and increase max_depth

+++ analyzer/src/a_star.rs
@@ -10,10 +10,10 @@ pub trait SortableCtxWrapper<T: Ctx, P: Ord>: std::fmt::Debug + Ord {
     fn ctx(&self) -> &ContextWrapper<T>;
     fn copy_update(&self, newctx: ContextWrapper<T>, score: P) -> Self;
     fn new_incr(&self, newctx: ContextWrapper<T>, score: P) -> Self;
-    fn should_keep(&self, _max_depth: i8) -> bool {
+    fn should_keep(&self, _max_depth: usize) -> bool {
         true
     }
-    fn can_continue(&self, _max_depth: i8) -> bool {
+    fn can_continue(&self, _max_depth: usize) -> bool {
         true
     }
 }
@@ -35,7 +35,7 @@ pub struct ScoredCtxWithActionCounter<T: Ctx> {
     #[sort_by]
     pub(crate) score: u32,
     pub(crate) el: ContextWrapper<T>,
-    pub(crate) counter: i8,
+    pub(crate) counter: usize,
 }
 
 impl<T: Ctx> SortableCtxWrapper<T, u32> for ScoredCtxWithActionCounter<T> {
@@ -61,10 +61,10 @@ impl<T: Ctx> SortableCtxWrapper<T, u32> for ScoredCtxWithActionCounter<T> {
             counter: self.counter + 1,
         }
     }
-    fn should_keep(&self, max_depth: i8) -> bool {
+    fn should_keep(&self, max_depth: usize) -> bool {
         self.counter <= max_depth
     }
-    fn can_continue(&self, max_depth: i8) -> bool {
+    fn can_continue(&self, max_depth: usize) -> bool {
         self.counter < max_depth
     }
 }
